Added typing info and moving node-ical.js to index.js
